rename usecase
