+++ app/modules/CoreModule.scala
@@ -1,7 +1,7 @@
 package modules
 
 import com.google.inject.AbstractModule
-import dev.tchiba.sdmt.application.interactors.domainmodel.add.AddDomainModelInteractor
+import dev.tchiba.sdmt.application.interactors.domainmodel.add.CreateDomainModelInteractor
 import dev.tchiba.sdmt.application.interactors.domainmodel.update.UpdateDomainModelInteractor
 import dev.tchiba.sdmt.application.interactors.boundedContext.create.CreateBoundedContextInteractor
 import dev.tchiba.sdmt.application.interactors.boundedContext.update.UpdateBoundedContextInteractor
@@ -11,7 +11,7 @@ import dev.tchiba.sdmt.core.user.UserRepository
 import dev.tchiba.sdmt.infra.domainmodel.{JdbcDomainModelRepository, JdbcDomainModelValidator}
 import dev.tchiba.sdmt.infra.boundedContext.JdbcBoundedContextRepository
 import dev.tchiba.sdmt.infra.user.JdbcUserRepository
-import dev.tchiba.sdmt.usecase.domainmodel.add.AddDomainModelUseCase
+import dev.tchiba.sdmt.usecase.domainmodel.create.CreateDomainModelUseCase
 import dev.tchiba.sdmt.usecase.domainmodel.update.UpdateDomainModelUseCase
 import dev.tchiba.sdmt.usecase.boundedContext.create.CreateBoundedContextUseCase
 import dev.tchiba.sdmt.usecase.boundedContext.update.UpdateBoundedContextUseCase
@@ -24,7 +24,7 @@ class CoreModule extends AbstractModule with ScalaModule {
 
     bind[CreateBoundedContextUseCase].to[CreateBoundedContextInteractor]
     bind[UpdateBoundedContextUseCase].to[UpdateBoundedContextInteractor]
-    bind[AddDomainModelUseCase].to[AddDomainModelInteractor]
+    bind[CreateDomainModelUseCase].to[CreateDomainModelInteractor]
 
     bind[UpdateDomainModelUseCase].to[UpdateDomainModelInteractor]
     bind[DomainModelValidator].to[JdbcDomainModelValidator]
