+++ sdmt-application/src/main/scala/dev/tchiba/sdmt/application/interactors/domainmodel/add/CreateDomainModelInteractor.scala
@@ -2,18 +2,22 @@ package dev.tchiba.sdmt.application.interactors.domainmodel.add
 
 import dev.tchiba.sdmt.core.boundedContext.BoundedContextRepository
 import dev.tchiba.sdmt.core.domainmodel.{DomainModel, DomainModelRepository}
-import dev.tchiba.sdmt.usecase.domainmodel.add.{AddDomainModelInput, AddDomainModelOutput, AddDomainModelUseCase}
+import dev.tchiba.sdmt.usecase.domainmodel.create.{
+  CreateDomainModelInput,
+  CreateDomainModelOutput,
+  CreateDomainModelUseCase
+}
 
 import javax.inject.Inject
 
-class AddDomainModelInteractor @Inject() (
+class CreateDomainModelInteractor @Inject() (
     boundedContextRepository: BoundedContextRepository,
     domainModelRepository: DomainModelRepository
-) extends AddDomainModelUseCase {
+) extends CreateDomainModelUseCase {
 
-  override def handle(input: AddDomainModelInput): AddDomainModelOutput = {
+  override def handle(input: CreateDomainModelInput): CreateDomainModelOutput = {
     boundedContextRepository.findByAlias(input.boundedContextAlias) match {
-      case None => AddDomainModelOutput.NoSuchBoundedContext(input.boundedContextAlias)
+      case None => CreateDomainModelOutput.NoSuchBoundedContext(input.boundedContextAlias)
       case Some(context) =>
         val newDomainModel = DomainModel.create(
           boundedCOntextId = context.id,
@@ -22,10 +26,10 @@ class AddDomainModelInteractor @Inject() (
           specification = input.specification
         )
         domainModelRepository.findByEnglishName(newDomainModel.englishName, newDomainModel.boundedContextId) match {
-          case Some(_) => AddDomainModelOutput.ConflictEnglishName(newDomainModel.englishName)
+          case Some(_) => CreateDomainModelOutput.ConflictEnglishName(newDomainModel.englishName)
           case None =>
             domainModelRepository.insert(newDomainModel)
-            AddDomainModelOutput.Success(newDomainModel)
+            CreateDomainModelOutput.Success(newDomainModel)
         }
     }
   }
