--- sdmt-usecase/src/main/scala/dev/tchiba/sdmt/usecase/domainmodel/add/AddDomainModelUseCase.scala
@@ -1,5 +1,0 @@
-package dev.tchiba.sdmt.usecase.domainmodel.add
-
-import dev.tchiba.sdmt.usecase.UseCase
-
-abstract class AddDomainModelUseCase extends UseCase[AddDomainModelInput, AddDomainModelOutput]
