+++ sdmt-usecase/src/main/scala/dev/tchiba/sdmt/usecase/domainmodel/create/CreateDomainModelInput.scala
@@ -1,11 +1,11 @@
-package dev.tchiba.sdmt.usecase.domainmodel.add
+package dev.tchiba.sdmt.usecase.domainmodel.create
 
 import dev.tchiba.sdmt.core.boundedContext.BoundedContextAlias
 import dev.tchiba.sdmt.usecase.Input
 
-final case class AddDomainModelInput(
+final case class CreateDomainModelInput(
     boundedContextAlias: BoundedContextAlias,
     japaneseName: String,
     englishName: String,
     specification: String
-) extends Input[AddDomainModelOutput]
+) extends Input[CreateDomainModelOutput]
