+++ sdmt-usecase/src/main/scala/dev/tchiba/sdmt/usecase/domainmodel/create/CreateDomainModelUseCase.scala
@@ -1,0 +1,5 @@
+package dev.tchiba.sdmt.usecase.domainmodel.create
+
+import dev.tchiba.sdmt.usecase.UseCase
+
+abstract class CreateDomainModelUseCase extends UseCase[CreateDomainModelInput, CreateDomainModelOutput]
